WIP mid-refactor
