--- fetchSync/actionTypes.js
@@ -1,17 +1,0 @@
-'use strict'
-
-export const ADD_SYNC = 'ADD_SYNC'
-export const ADD_SYNCS = 'ADD_SYNCS'
-
-export const REMOVE_SYNC = 'REMOVE_SYNC'
-export const REMOVE_ALL_SYNCS = 'REMOVE_ALL_SYNCS'
-
-export const SET_SERVICE_WORKER = 'SET_SERVICE_WORKER'
-export const SET_COMMS_OPEN = 'SET_COMMS_OPEN'
-
-export const Requests = {
-  OPEN_COMMS: 'OPEN_COMMS',
-  REGISTER_SYNC: 'REGISTER_SYNC',
-  CANCEL_SYNC: 'CANCEL_SYNC',
-  CANCEL_ALL: 'CANCEL_ALL'
-}
