--- fetchSync/constants.js
@@ -1,8 +1,0 @@
-'use strict'
-
-export const SyncIdPrefix = 'FETCH_SYNC__'
-
-export const CommsChannelStatus = {
-  OPEN: 1,
-  CLOSED: 0
-}
