+++ src/Messages.js
@@ -1,0 +1,5 @@
+export const OPEN_COMMS = 'OPEN_COMMS'
+export const REGISTER_SYNC = 'REGISTER_SYNC'
+export const CANCEL_SYNC = 'CANCEL_SYNC'
+export const CANCEL_ALL_SYNCS = 'CANCEL_ALL_SYNCS'
+export const SYNC_RESULT = 'SYNC_RESULT'
